// src/types/formData.ts
export interface FormDataType {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  linkedin: string;
  portfolio: string;
  company: string;
  industry: string;
  description: string;
  serviceType: string;
  features: string;
  timeline: string;
  budget: string;
  additionalInfo: string;
}